refactor(identity): fail fast on invalid input before BCrypt hashing (#10)
- Extract reusable validation helpers into `Preconditions`
- Centralize null/blank checks and email normalization logic
- Reuse `Preconditions` across `User` and `IdentityService`
- Validate and normalize email before invoking BCrypt hashing
- Avoid unnecessary password hashing work for malformed input
- Simplify `User` constructor by removing duplicated inline validation code
- Preserve domain invariants inside the `User` constructor
